+++ scripts/translate.ts
@@ -102,7 +102,6 @@ async function main() {
   await completeTranslationsFor('toolbar', keysToUpdate, deleteKeys);
   await completeTranslationsFor('seelenweg', keysToUpdate, deleteKeys);
   await completeTranslationsFor('settings', keysToUpdate, deleteKeys);
-  await completeTranslationsFor('update', keysToUpdate, deleteKeys);
   await completeTranslationsFor('seelen_rofi', keysToUpdate, deleteKeys);
 }