+++ src/apps/settings/components/header/index.tsx
@@ -1,19 +1,27 @@
+import { invoke } from '@tauri-apps/api/core';
 import { getCurrentWebviewWindow } from '@tauri-apps/api/webviewWindow';
 import { relaunch } from '@tauri-apps/plugin-process';
-import { Button, Tooltip } from 'antd';
+import { check as getUpdate, Update } from '@tauri-apps/plugin-updater';
+import { Badge, Button, Tooltip } from 'antd';
+import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
+import { SeelenCommand } from 'seelen-core';
 
 import { SaveStore } from '../../modules/shared/store/infra';
 import { useAppSelector } from '../../modules/shared/utils/infra';
 
 import { RootActions } from '../../modules/shared/store/app/reducer';
 import { RootSelectors } from '../../modules/shared/store/app/selectors';
+import { wasInstalledUsingMSIX } from 'src/apps/shared';
+import { Icon } from 'src/apps/shared/components/Icon';
 
 import { RouteExtraInfo } from '../navigation/routes';
 import cs from './index.module.css';
 
 export const Header = () => {
+  let [update, setUpdate] = useState<Update | null>(null);
+
   let route = useAppSelector(RootSelectors.route);
   let hasChanges = useAppSelector(RootSelectors.toBeSaved);
   let shouldRestart = useAppSelector(RootSelectors.toBeRestarted);
@@ -21,6 +29,21 @@ export const Header = () => {
   const { t } = useTranslation();
 
   const dispatch = useDispatch();
+
+  useEffect(() => {
+    async function checkUpdate() {
+      let isDev = await invoke<boolean>(SeelenCommand.IsDevMode);
+      let isMsix = await wasInstalledUsingMSIX();
+      if (!isDev && !isMsix) {
+        let update = await getUpdate();
+        if (update) {
+          await update.download();
+          setUpdate(update);
+        }
+      }
+    }
+    checkUpdate();
+  }, []);
 
   const cancelChanges = () => {
     dispatch(RootActions.restoreToLastLoaded());
@@ -48,16 +71,26 @@ export const Header = () => {
           </Tooltip>
         )}
       </div>
-      <div>
+      <div className={cs.actions}>
+        {update && (
+          <Tooltip title={t('update.available')}>
+            <Button type="text" onClick={() => update?.install()}>
+              <Badge dot>
+                <Icon iconName="TbDownload" />
+              </Badge>
+            </Button>
+          </Tooltip>
+        )}
         <Button
+          style={{ minWidth: 60 }}
           children={t('cancel')}
           type="default"
           danger
           disabled={!hasChanges}
           onClick={cancelChanges}
         />
-        {'  '}
         <Button
+          style={{ minWidth: 60 }}
           children={hasChanges ? saveLabel : t('close')}
           type="primary"
           danger={!hasChanges}